GoogleCalendar API Completion.

+++ WebSimplify/WebSimplify/DataAccess/SqlDbGoogle.cs
@@ -5,6 +5,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Web;
+using WebSimplify.Data;
 
 namespace WebSimplify
 {
@@ -12,6 +13,21 @@ namespace WebSimplify
     {
         public SqlDbGoogle(string _connectionString) : base(new SynnSqlDataProvider(_connectionString))
         {
+        }
+
+       
+        int appUserId { get; set; }
+
+        int IDbGoogle.AppUserId
+        {
+            get
+            {
+                return appUserId;
+            }
+            set
+            {
+                appUserId = value;
+            }
         }
 
         public void ClearData()
@@ -31,8 +47,8 @@ namespace WebSimplify
         public void DeleteStoredKey(string key)
         {
             CheckKey(key);
-            SetSqlFormat("DELETE FROM {0} WHERE userid = ?", SynnDataProvider.TableNames.GoogleTokens);
-            SetParameters(key);
+            SetSqlFormat("DELETE FROM {0} WHERE userid = ? and appuserid = ?", SynnDataProvider.TableNames.GoogleTokens);
+            SetParameters(key, appUserId);
             ExecuteSql();
         }
 
@@ -41,10 +57,20 @@ namespace WebSimplify
             CheckKey(key);
             SetSqlFormat("select credentials from {0}", SynnDataProvider.TableNames.GoogleTokens);
             ClearParameters();
-            
+
+            AddSqlWhereField("appuserid", appUserId);
             AddSqlWhereLikeField("userid", key, LikeSelectionStyle.AsIs);
 
-            return GetSingleRecordFirstValue().ToString();
+            try
+            {
+                var res = GetSingleRecordFirstValue();
+
+                return res == null ? string.Empty : res.ToString();
+            }
+            catch (Exception ex)
+            {
+                return string.Empty;
+            }
         }
 
         public void Upsert(string key, string serializedCredentials)
@@ -54,7 +80,11 @@ namespace WebSimplify
             var userCreds = GetUserCredentialsByKey(key);
             if (!string.IsNullOrEmpty(userCreds))
             {
-                SqlItemList where = new SqlItemList { new SqlItem("userid", key) };
+                SqlItemList where = new SqlItemList
+                {
+                    new SqlItem("userid", key),
+                    new SqlItem("appuserid", appUserId)
+                };
                 SqlItemList items = new SqlItemList { new SqlItem("credentials", serializedCredentials) };
                 SetUpdateSql(SynnDataProvider.TableNames.GoogleTokens, items, where);
             }
@@ -62,20 +92,12 @@ namespace WebSimplify
             {
                 SqlItemList items = new SqlItemList();
                 items.Add(new SqlItem("userid", key));
+                items.Add(new SqlItem("appuserid", appUserId));
                 items.Add(new SqlItem("credentials", serializedCredentials));
                 SetInsertIntoSql(SynnDataProvider.TableNames.GoogleTokens, items);
             }
             ExecuteSql();
         }
-
-        public string GetCredentialsJsonString(int userId)
-        {
-            SetSqlFormat("select credentials from {0}", SynnDataProvider.TableNames.GoogleAPICredentials);
-            ClearParameters();
-
-            AddSqlWhereField("userid", userId);
-
-            return GetSingleRecordFirstValue().ToString();
-        }
+        
     }
 }
